Add server node.js file
Very basic server code outline. Needs mysql and node.js to run

Right now it is a command line interface that lets you create a database through the command line using the create() function, just as a proof of concept of what's possible. The idea is that it eventually will let you use the website "normally" but through the command line.

To do:
-Add/implement functions mentioned above
-Add html parser + sender
-Add login functionality